FLUX.2/requirements.txt
torch==2.8.0
torchvision==0.23.0
diffusers
transformers
accelerate
safetensors
huggingface_hub
Pillow
numpy
matplotlib
PyYAML
tqdm
seaborn
lpips
